Go back to .contains() for engine names
This broke for cases in CCRL because of overlapping engine names.
Theoretically that could still be an issue, but the bigger issue is that
in TCEC, the engine names contain more information (e.g. SHAs) and are
harder to normalize.

There'll be a better solution for this later, but for now this will
help.